requirements.txt
ultralytics==8.3.234
onnxruntime-gpu
flask
gunicorn
numpy
opencv-python-headless
pillow